+++ packages/components/src/utils/isRefObject.ts
@@ -1,0 +1,3 @@
+export const isRefObject = (
+  ref: React.ForwardedRef<any>
+): ref is React.MutableRefObject<any> => ref !== null && "current" in ref