+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.4.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@16.4.2...@kaizen/component-library@16.4.3) (2023-03-14)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
 ## [16.4.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@16.4.1...@kaizen/component-library@16.4.2) (2023-03-02)